feat: Implement ResponseEntity for create and delete routes

+++ src/main/java/br/com/guilhermesalvador/uaheroacademyapi/controller/CourseController.java
@@ -2,6 +2,8 @@ package br.com.guilhermesalvador.uaheroacademyapi.controller;
 
 import br.com.guilhermesalvador.uaheroacademyapi.dto.CourseDTO;
 import br.com.guilhermesalvador.uaheroacademyapi.service.CourseService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,8 +19,10 @@ public class CourseController {
     }
 
     @PostMapping
-    public CourseDTO create(@RequestBody CourseDTO course) {
-        return courseService.create(course);
+    public ResponseEntity<String> create(@RequestBody CourseDTO course) {
+        CourseDTO createdCourse = courseService.create(course);
+
+        return ResponseEntity.status(HttpStatus.CREATED).body("Course created successfully");
     }
 
     @GetMapping
@@ -32,8 +36,13 @@ public class CourseController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Long id) {
-        courseService.deleteById(id);
+    public ResponseEntity<String> deleteById(@PathVariable Long id) {
+        if (courseService.findById(id) != null) {
+            courseService.deleteById(id);
+            return ResponseEntity.ok("Course deleted successfully");
+        }
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Course with id " + id + " not found");
     }
 
     @PutMapping("/{id}")
